JB-663 - Search doesn't work on mobile
